worker: refactor upload
