+++ worker/uploader.go
@@ -1,0 +1,265 @@
+package worker
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"math"
+	"sync"
+	"time"
+
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/trace"
+	rhpv2 "go.sia.tech/core/rhp/v2"
+	"go.sia.tech/core/types"
+	"go.sia.tech/renterd/api"
+	"go.sia.tech/renterd/stats"
+)
+
+type (
+	uploader struct {
+		b Bus
+
+		hk              types.PublicKey
+		siamuxAddr      string
+		signalNewUpload chan struct{}
+		shutdownCtx     context.Context
+
+		mu        sync.Mutex
+		bh        uint64
+		endHeight uint64
+		fcid      types.FileContractID
+		host      hostV3
+		queue     []*sectorUploadReq
+
+		// stats related field
+		consecutiveFailures uint64
+		lastRecompute       time.Time
+
+		statsSectorUploadEstimateInMS    *stats.DataPoints
+		statsSectorUploadSpeedBytesPerMS *stats.DataPoints
+	}
+)
+
+func (u *uploader) BlockHeight() uint64 {
+	u.mu.Lock()
+	defer u.mu.Unlock()
+	return u.bh
+}
+
+func (u *uploader) ContractID() types.FileContractID {
+	u.mu.Lock()
+	defer u.mu.Unlock()
+	return u.fcid
+}
+
+func (u *uploader) Healthy() bool {
+	u.mu.Lock()
+	defer u.mu.Unlock()
+	return u.consecutiveFailures == 0
+}
+
+func (u *uploader) Renew(hp hostProvider, c api.ContractMetadata, bh uint64) {
+	u.mu.Lock()
+	defer u.mu.Unlock()
+
+	u.bh = bh
+	u.host = hp.newHostV3(c.ID, c.HostKey, c.SiamuxAddr)
+	u.fcid = c.ID
+	u.siamuxAddr = c.SiamuxAddr
+	u.endHeight = c.WindowEnd
+}
+
+func (u *uploader) SignalWork() {
+	select {
+	case u.signalNewUpload <- struct{}{}:
+	default:
+	}
+}
+
+func (u *uploader) Start(hp hostProvider, rl revisionLocker) {
+outer:
+	for {
+		// wait for work
+		select {
+		case <-u.signalNewUpload:
+		case <-u.shutdownCtx.Done():
+			return
+		}
+
+		for {
+			// check if we are stopped
+			select {
+			case <-u.shutdownCtx.Done():
+				return
+			default:
+			}
+
+			// pop the next upload req
+			req := u.pop()
+			if req == nil {
+				continue outer
+			}
+
+			// skip if upload is done
+			if req.done() {
+				continue
+			}
+
+			// execute it
+			var root types.Hash256
+			start := time.Now()
+			fcid := u.ContractID()
+			err := rl.withRevision(req.sector.ctx, defaultRevisionFetchTimeout, fcid, u.hk, u.siamuxAddr, req.lockPriority, u.BlockHeight(), func(rev types.FileContractRevision) error {
+				if rev.RevisionNumber == math.MaxUint64 {
+					return errMaxRevisionReached
+				}
+
+				var err error
+				root, err = u.execute(req, rev)
+				return err
+			})
+
+			// the uploader's contract got renewed, requeue the request
+			if errors.Is(err, errMaxRevisionReached) {
+				u.enqueue(req)
+				continue outer
+			}
+
+			// send the response
+			if err != nil {
+				req.fail(err)
+			} else {
+				req.succeed(root)
+			}
+
+			// track the error, ignore gracefully closed streams and canceled overdrives
+			canceledOverdrive := req.done() && req.overdrive && err != nil
+			if !canceledOverdrive && !isClosedStream(err) {
+				u.trackSectorUpload(err, time.Since(start))
+			}
+		}
+	}
+}
+
+func (u *uploader) Stop() {
+	for {
+		upload := u.pop()
+		if upload == nil {
+			break
+		}
+		if !upload.done() {
+			upload.fail(errors.New("uploader stopped"))
+		}
+	}
+}
+
+func (u *uploader) UpdateBlockHeight(bh uint64) {
+	u.mu.Lock()
+	defer u.mu.Unlock()
+	u.bh = bh
+}
+
+func (u *uploader) enqueue(req *sectorUploadReq) {
+	// trace the request
+	span := trace.SpanFromContext(req.sector.ctx)
+	span.SetAttributes(attribute.Stringer("hk", u.hk))
+	span.AddEvent("enqueued")
+
+	// decorate the request
+	req.fcid = u.ContractID()
+	req.hk = u.hk
+
+	// enqueue the request
+	u.mu.Lock()
+	u.queue = append(u.queue, req)
+	u.mu.Unlock()
+
+	// signal there's work
+	u.SignalWork()
+}
+
+func (u *uploader) estimate() float64 {
+	u.mu.Lock()
+	defer u.mu.Unlock()
+
+	// fetch estimated duration per sector
+	estimateP90 := u.statsSectorUploadEstimateInMS.P90()
+	if estimateP90 == 0 {
+		estimateP90 = 1
+	}
+
+	// calculate estimated time
+	numSectors := float64(len(u.queue) + 1)
+	return numSectors * estimateP90
+}
+
+func (u *uploader) execute(req *sectorUploadReq, rev types.FileContractRevision) (types.Hash256, error) {
+	u.mu.Lock()
+	host := u.host
+	fcid := u.fcid
+	u.mu.Unlock()
+
+	// fetch span from context
+	span := trace.SpanFromContext(req.sector.ctx)
+	span.AddEvent("execute")
+
+	// update the bus
+	if err := u.b.AddUploadingSector(req.sector.ctx, req.uploadID, fcid, req.sector.root); err != nil {
+		return types.Hash256{}, fmt.Errorf("failed to add uploading sector to contract %v, err: %v", fcid, err)
+	}
+
+	// upload the sector
+	start := time.Now()
+	root, err := host.UploadSector(req.sector.ctx, req.sector.data, rev)
+	if err != nil {
+		return types.Hash256{}, err
+	}
+
+	// update span
+	elapsed := time.Since(start)
+	span.SetAttributes(attribute.Int64("duration", elapsed.Milliseconds()))
+	span.RecordError(err)
+	span.End()
+
+	return root, nil
+}
+
+func (u *uploader) pop() *sectorUploadReq {
+	u.mu.Lock()
+	defer u.mu.Unlock()
+
+	if len(u.queue) > 0 {
+		j := u.queue[0]
+		u.queue[0] = nil
+		u.queue = u.queue[1:]
+		return j
+	}
+	return nil
+}
+
+func (u *uploader) trackSectorUpload(err error, d time.Duration) {
+	u.mu.Lock()
+	defer u.mu.Unlock()
+	if err != nil {
+		u.consecutiveFailures++
+		u.statsSectorUploadEstimateInMS.Track(float64(time.Hour.Milliseconds()))
+	} else {
+		ms := d.Milliseconds()
+		u.consecutiveFailures = 0
+		u.statsSectorUploadEstimateInMS.Track(float64(ms))                       // duration in ms
+		u.statsSectorUploadSpeedBytesPerMS.Track(float64(rhpv2.SectorSize / ms)) // bytes per ms
+	}
+}
+
+func (u *uploader) tryRecomputeStats() {
+	u.mu.Lock()
+	defer u.mu.Unlock()
+	if time.Since(u.lastRecompute) < statsRecomputeMinInterval {
+		return
+	}
+
+	u.lastRecompute = time.Now()
+	u.statsSectorUploadEstimateInMS.Recompute()
+	u.statsSectorUploadSpeedBytesPerMS.Recompute()
+}
