+++ worker/worker.go
@@ -107,12 +107,6 @@ type (
 	}
 )
 
-type ContractLocker interface {
-	AcquireContract(ctx context.Context, fcid types.FileContractID, priority int, d time.Duration) (lockID uint64, err error)
-	KeepaliveContract(ctx context.Context, fcid types.FileContractID, lockID uint64, d time.Duration) (err error)
-	ReleaseContract(ctx context.Context, fcid types.FileContractID, lockID uint64) (err error)
-}
-
 // A Bus is the source of truth within a renterd system.
 type Bus interface {
 	alerts.Alerter
@@ -311,20 +305,6 @@ func (w *worker) newHostV3(contractID types.FileContractID, hostKey types.Public
 		transportPool:            w.transportPoolV3,
 		priceTables:              w.priceTables,
 	}
-}
-
-func (w *worker) withContractLock(ctx context.Context, fcid types.FileContractID, priority int, fn func() error) error {
-	contractLock, err := w.acquireContractLock(ctx, fcid, priority)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		releaseCtx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-		_ = contractLock.Release(releaseCtx)
-		cancel()
-	}()
-
-	return fn()
 }
 
 func (w *worker) withRevision(ctx context.Context, fetchTimeout time.Duration, contractID types.FileContractID, hk types.PublicKey, siamuxAddr string, lockPriority int, blockHeight uint64, fn func(rev types.FileContractRevision) error) error {
@@ -1283,18 +1263,6 @@ func (w *worker) multipartUploadHandlerPUT(jc jape.Context) {
 	jc.ResponseWriter.Header().Set("ETag", api.FormatETag(eTag))
 }
 
-func encryptPartialSlab(data []byte, key object.EncryptionKey, minShards, totalShards uint8) [][]byte {
-	slab := object.Slab{
-		Key:       key,
-		MinShards: minShards,
-		Shards:    make([]object.Sector, totalShards),
-	}
-	encodedShards := make([][]byte, totalShards)
-	slab.Encode(data, encodedShards)
-	slab.Encrypt(encodedShards)
-	return encodedShards
-}
-
 func (w *worker) objectsHandlerDELETE(jc jape.Context) {
 	var batch bool
 	if jc.DecodeForm("batch", &batch) != nil {
@@ -1342,10 +1310,6 @@ func (w *worker) rhpContractsHandlerGET(jc jape.Context) {
 		resp.Error = errs.Error()
 	}
 	jc.Encode(resp)
-}
-
-func preparePayment(accountKey types.PrivateKey, amt types.Currency, blockHeight uint64) rhpv3.PayByEphemeralAccountRequest {
-	return rhpv3.PayByEphemeralAccount(rhpv3.Account(accountKey.PublicKey()), amt, blockHeight+6, accountKey) // 1 hour valid
 }
 
 func (w *worker) idHandlerGET(jc jape.Context) {
@@ -1480,91 +1444,6 @@ func (w *worker) Shutdown(_ context.Context) error {
 	// Stop the uploader.
 	w.uploadManager.Stop()
 	return nil
-}
-
-type contractLock struct {
-	lockID uint64
-	fcid   types.FileContractID
-	d      time.Duration
-	locker ContractLocker
-	logger *zap.SugaredLogger
-
-	stopCtx       context.Context
-	stopCtxCancel context.CancelFunc
-	stopWG        sync.WaitGroup
-}
-
-func newContractLock(fcid types.FileContractID, lockID uint64, d time.Duration, locker ContractLocker, logger *zap.SugaredLogger) *contractLock {
-	ctx, cancel := context.WithCancel(context.Background())
-	cl := &contractLock{
-		lockID: lockID,
-		fcid:   fcid,
-		d:      d,
-		locker: locker,
-		logger: logger,
-
-		stopCtx:       ctx,
-		stopCtxCancel: cancel,
-	}
-	cl.stopWG.Add(1)
-	go func() {
-		cl.keepaliveLoop()
-		cl.stopWG.Done()
-	}()
-	return cl
-}
-
-func (cl *contractLock) Release(ctx context.Context) error {
-	// Stop background loop.
-	cl.stopCtxCancel()
-	cl.stopWG.Wait()
-
-	// Release the contract.
-	return cl.locker.ReleaseContract(ctx, cl.fcid, cl.lockID)
-}
-
-func (cl *contractLock) keepaliveLoop() {
-	// Create ticker for 20% of the lock duration.
-	start := time.Now()
-	var lastUpdate time.Time
-	tickDuration := cl.d / 5
-	t := time.NewTicker(tickDuration)
-
-	// Cleanup
-	defer func() {
-		t.Stop()
-		select {
-		case <-t.C:
-		default:
-		}
-	}()
-
-	// Loop until stopped.
-	for {
-		select {
-		case <-cl.stopCtx.Done():
-			return // released
-		case <-t.C:
-		}
-		if err := cl.locker.KeepaliveContract(cl.stopCtx, cl.fcid, cl.lockID, cl.d); err != nil && !errors.Is(err, context.Canceled) {
-			cl.logger.Errorw(fmt.Sprintf("failed to send keepalive: %v", err),
-				"contract", cl.fcid,
-				"lockID", cl.lockID,
-				"loopStart", start,
-				"timeSinceLastUpdate", time.Since(lastUpdate),
-				"tickDuration", tickDuration)
-			return
-		}
-		lastUpdate = time.Now()
-	}
-}
-
-func (w *worker) acquireContractLock(ctx context.Context, fcid types.FileContractID, priority int) (_ revisionUnlocker, err error) {
-	lockID, err := w.bus.AcquireContract(ctx, fcid, priority, w.contractLockingDuration)
-	if err != nil {
-		return nil, err
-	}
-	return newContractLock(fcid, lockID, w.contractLockingDuration, w.bus, w.logger), nil
 }
 
 func (w *worker) scanHost(ctx context.Context, hostKey types.PublicKey, hostIP string) (settings rhpv2.HostSettings, pt rhpv3.HostPriceTable, elapsed time.Duration, err error) {
